visuaaaaaaaaaals
but you probably can't see them! try to run the v1 files in the folder i
added, but chances are it wont work out of the box because export to mac
osx is broken in processing 2. buuuut if you install at processing.org
you can open up the .pde file for sure! plus processing is amazing you
would love it